Crm updates

+++ app/Http/Controllers/ProductController.php
@@ -17,6 +17,10 @@ use DB;
 use Maatwebsite\Excel\Facades\Excel;
 use App\Exports\ProductsExport;
 use App\Imports\ProductsImport;
+use App\Imports\BuyingPriceImport;
+use App\Models\Currency;
+use App\Models\BuyingPrice;
+use Carbon\Carbon;
 
 class ProductController extends Controller
 {
@@ -25,13 +29,13 @@ class ProductController extends Controller
    *
    * @return \Illuminate\Http\Response
    */
-  public function index(Request $request, ProductService $productService,SupplierService $supplierService)
+  public function index(Request $request, ProductService $productService, SupplierService $supplierService)
   {
 
     $input = $request->all();
     $products = $productService->getAllProduct($input);
     $suppliers = $supplierService->getAllSupplier();
-    return view('products.index', compact('products','suppliers'));
+    return view('products.index', compact('products', 'suppliers'));
   }
 
   /**
@@ -48,9 +52,10 @@ class ProductController extends Controller
     $suppliers = $supplierService->getAllSupplier();
     $divisions = $divisionService->getDivisionList();
     $managers = $productService->employeesList();
-    $paymentTerms = PaymentTerm::where('parent_id',0)->orderByDesc('isdefault')->get();
-    $currencies = DB::table('currency')->get();
+    $paymentTerms = PaymentTerm::where('parent_id', 0)->orderByDesc('isdefault')->get();
+    $currencies = Currency::orderBy('code', 'asc')->get();
     $currencyConversions = DB::table('currency_conversion')->get();
+
     return view('products.create',  compact(
       'suppliers',
       'divisions',
@@ -58,6 +63,7 @@ class ProductController extends Controller
       'paymentTerms',
       'currencyConversions',
       'currencies'
+
     ));
   }
 
@@ -118,11 +124,21 @@ class ProductController extends Controller
     $suppliers = $supplierService->getAllSupplier();
     $divisions = $divisionService->getDivisionList();
     $managers = $productService->employeesList();
-    $paymentTerms = PaymentTerm::where('parent_id',0)->get();
-    $currencies = DB::table('currency')->get();
+    $paymentTerms = PaymentTerm::where('parent_id', 0)->get();
+    $currencies = Currency::orderBy('code', 'asc')->get();
     $currencyConversions = DB::table('currency_conversion')->get();
     $productPriceHistory = ProductPriceHistory::where('product_id', $id)
       ->get();
+    $purchasePriceHistory = BuyingPrice::where('product_id', $product->id)->get();
+
+    // check month diff
+    if (!empty($product->buyingPrice) && isset($product->buyingPrice[0])) {
+      $valid_from = $product->buyingPrice[0]->validity_from;
+      $valid_to = $product->buyingPrice[0]->validity_to;
+      $date1 = new Carbon($valid_from);    // start date
+      $date2 = new Carbon($valid_to);  // end date
+      $product->buyingPrice[0]->buyingPriceMonth = $date1->diffInMonths($date2);
+    }
     return view('products.edit',  compact(
       'product',
       'suppliers',
@@ -131,7 +147,8 @@ class ProductController extends Controller
       'paymentTerms',
       'currencyConversions',
       'currencies',
-      'productPriceHistory'
+      'productPriceHistory',
+      'purchasePriceHistory'
     ));
   }
 
@@ -154,7 +171,6 @@ class ProductController extends Controller
     }
     if (isset($request->remove_image)) {
       ($product->image_url) ? $productService->deleteImage($product->image_url) : '';
-
     }
 
     $productService->updateProduct($product, $input, $image_url);
@@ -172,7 +188,8 @@ class ProductController extends Controller
   {
     $product = $productService->getProduct($id);
 
-    $productService->deleteImage($product->image_url);
+    if ($product->image_url)
+      $productService->deleteImage($product->image_url);
 
     $productService->DeleteProduct($product);
 
@@ -188,7 +205,8 @@ class ProductController extends Controller
     $suppliers = $supplierService->getAllSupplier();
     $divisions = $divisionService->getDivisionList();
     $managers = $productService->employeesList();
-    return view('products.product-import', compact('suppliers', 'divisions', 'managers'));
+    $currencies = Currency::all();
+    return view('products.product-import', compact('suppliers', 'divisions', 'managers', 'currencies'));
   }
   public function importSave(Request $request)
   {
@@ -205,5 +223,56 @@ class ProductController extends Controller
   public function downloadExcelTemplate()
   {
     return Excel::download(new ProductsExport, 'products_template.xlsx');
+  }
+
+  public function buyingPriceSave(Request $request)
+  {
+    $brandId = $request->input('brand_id');
+    $currency = $request->input('currency');
+
+    $file = $request->file('files');
+    try {
+      Excel::import(new BuyingPriceImport($brandId, $currency), $file);
+      return redirect()->back()->with('success', 'Data imported successfully!');
+    } catch (\Exception $e) {
+      return redirect()->back()->with('error', 'Error importing file: ' . $e->getMessage());
+    }
+  }
+
+  public function saveAjaxBuyingPrice(
+    Request $request,
+    ProductService $productService
+  ) {
+
+    $data = $request->all();
+
+    $buyingPrice = $productService->createBuyingPrice($data);
+
+    return response()->json(['success' => 'Buying Price created successfully', 'data' => $buyingPrice]);
+  }
+
+  public function buyingPriceByProduct(
+    Request $request,
+    ProductService $productService
+  ) {
+
+    $data = $request->all();
+    $id = $data['product_id'];
+
+    $buying_price = $productService->buyingPriceByProduct($id);
+    $buyingPrice = 0;
+    if (!empty($buying_price)) {
+      $bprice = $buying_price->buying_price;
+      $bcurrency = $buying_price->buying_currency;
+
+      $currency_rate = DB::table('currency_conversion')->where('currency', $bcurrency)->first();
+      if ($currency_rate) {
+        $buyingPrice = $bprice * $currency_rate->standard_rate;
+        $buyingPrice = number_format((float)$buyingPrice, 2, '.', '');
+      }
+    }
+
+
+    return response()->json(['success' => 'Buying Price fetch successfully', 'data' => $buyingPrice]);
   }
 }
